requirements.txt
# Django 核心
Django>=6.0
djangorestframework>=3.15
django-cors-headers>=4.3

# 数据库
# SQLite 已内置，无需额外安装

# AI 相关
langchain>=0.3.0
langchain-community>=0.3.0
chromadb>=0.5.0
sentence-transformers>=3.0.0  # bge-m3 嵌入模型

# 智谱 AI（OpenAI 兼容）
openai>=1.50.0  # 智谱使用 OpenAI 兼容接口

# 文档处理
python-docx>=1.1.0
PyPDF2>=3.0.0
markdown>=3.7

# 部署
gunicorn>=23.0.0
whitenoise>=6.7.0

# 工具
python-dotenv>=1.0.0
requests>=2.32.0
